Manual rebased routing branch to v1dev source SHA-1: c29e0a1456816015e8f80771d5dc0c12877551e9

+++ src/knx/bau2920.cpp
@@ -97,7 +97,7 @@ InterfaceObject* Bau2920::getInterfaceObject(uint8_t idx)
     }
 }
 
-InterfaceObject* Bau2920::getInterfaceObject(ObjectType objectType, uint8_t objectInstance)
+InterfaceObject* Bau2920::getInterfaceObject(ObjectType objectType, uint16_t objectInstance)
 {
     // We do not use it right now. 
     // Required for coupler mode as there are multiple router objects for example
